principia-semantica/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "principia-semantica"
version = "0.1.0"
authors = [
  { name="Your Name", email="your@email.com" },
]
description = "A framework for the spectral geometry of meaning and thematic decomposition."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "numpy",
    "scipy",
    "scikit-learn",
    "pandas",
    "umap-learn",
    "matplotlib",
    "jupyterlab",
]

[project.urls]
Homepage = "https://github.com/your-username/principia-semantica"
Issues = "https://github.com/your-username/principia-semantica/issues"